fix(tooltip): fix candlestick tooltip display with xs option
- Make handling candlestick type shape in .findClosest()
- Move .isWithinBar() from bar.ts -> shape.ts

Ref #2434
